@@ -1,22 +1,27 @@
 name: grinder
 # This version must be updated in tandem with the lib/src/cli.dart version.
-version: 0.8.6
+version: 0.8.7-nullsafety.1
 description: Grinder is a task runner for Dart, helping to define and automate common project workflows.
 
 homepage: https://github.com/google/grinder.dart
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  cli_util: '>=0.1.2 <3.0.0'
-  glob: ^1.0.0
-  meta: ^1.1.7
-  path: ^1.0.0
+  cli_util: '>=0.3.0-nullsafety <1.0.0'
+  glob: ^2.0.0-nullsafety
+  meta: ^1.3.0-nullsafety
+  path: ^1.8.0-nullsafety
+  collection: ^1.15.0-nullsafety
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.0.0
+  pedantic: ^1.10.0-nullsafety
+  test: ^1.16.0-nullsafety
 
 executables:
   grind: grinder
+
+# Work around dart-lang/test#1400
+dependency_overrides:
+  glob: ^2.0.0-nullsafety
