@@ -16,19 +16,16 @@ import 'task_discovery/good_tasks.dart' as good;
 
 void main() {
   // Libs which contains annotated tasks (imported above).
-  LibraryMirror goodLib;
-  LibraryMirror badLib;
-  LibraryMirror externalLib;
-  externalLib;
+  late LibraryMirror goodLib;
+  late LibraryMirror badLib;
 
   // Dummy calls to avoid "unused import" warnings.
   good.variable;
   bad.dependsNonExported;
   external_tasks.shownVariable;
-  externalLib;
 
-  TaskDiscovery discoveryGood;
-  TaskDiscovery discoveryBad;
+  late TaskDiscovery discoveryGood;
+  late TaskDiscovery discoveryBad;
   setUp(() {
     LibraryMirror getLib(Symbol name) => currentMirrorSystem()
         .libraries
@@ -37,7 +34,6 @@ void main() {
 
     goodLib = getLib(#grinder.test.task_discovery.good_tasks);
     badLib = getLib(#grinder.test.task_discovery.bad_tasks);
-    externalLib = getLib(#grinder.test.task_discovery.external_tasks);
     discoveryGood = TaskDiscovery(goodLib);
     discoveryBad = TaskDiscovery(badLib);
   });
@@ -45,14 +41,14 @@ void main() {
   group('discoverDeclaration', () {
     test('should set cache', () {
       final cache = <DeclarationMirror, AnnotatedTask>{};
-      final methodDecl = goodLib.declarations[#method];
-      var annotated = discoveryGood.discoverDeclaration(methodDecl, cache);
+      final methodDecl = goodLib.declarations[#method]!;
+      var annotated = discoveryGood.discoverDeclaration(methodDecl, cache)!;
       expect(annotated.isDefault, isFalse);
       expect(cache, {methodDecl: annotated});
     });
 
     test('should get from cache', () {
-      var methodDecl = goodLib.declarations[#method];
+      var methodDecl = goodLib.declarations[#method]!;
       var annotated = AnnotatedTask(
           GrinderTask('method', taskFunction: nullTaskFunction), false);
       var cache = {methodDecl: annotated};
@@ -61,8 +57,8 @@ void main() {
     });
 
     test('should discover task from regular method', () {
-      var annotated =
-          discoveryGood.discoverDeclaration(goodLib.declarations[#method], {});
+      var annotated = discoveryGood
+          .discoverDeclaration(goodLib.declarations[#method]!, {})!;
       expect(annotated.isDefault, isFalse);
       var task = annotated.task;
       expect(task.name, 'method');
@@ -71,15 +67,15 @@ void main() {
 
     test('should discover task from variable', () {
       var annotated = discoveryGood
-          .discoverDeclaration(goodLib.declarations[#variable], {});
+          .discoverDeclaration(goodLib.declarations[#variable]!, {})!;
       var task = annotated.task;
       expect(task.name, 'variable');
       expect(task.depends, [TaskInvocation('method')]);
     });
 
     test('should discover task from getter', () {
-      var annotated =
-          discoveryGood.discoverDeclaration(goodLib.declarations[#getter], {});
+      var annotated = discoveryGood
+          .discoverDeclaration(goodLib.declarations[#getter]!, {})!;
       var task = annotated.task;
       expect(task.name, 'getter');
       expect(task.depends, [TaskInvocation('method')]);
@@ -87,13 +83,13 @@ void main() {
 
     test('should dasherize camel case task method', () {
       var annotated = discoveryGood
-          .discoverDeclaration(goodLib.declarations[#camelCase], {});
+          .discoverDeclaration(goodLib.declarations[#camelCase]!, {})!;
       expect(annotated.task.name, 'camel-case');
     });
 
     test('should discover a default task', () {
       var annotated =
-          discoveryGood.discoverDeclaration(goodLib.declarations[#def], {});
+          discoveryGood.discoverDeclaration(goodLib.declarations[#def]!, {})!;
       expect(annotated.isDefault, isTrue);
       var task = annotated.task;
       expect(task.name, 'def');
@@ -101,76 +97,76 @@ void main() {
     });
 
     test('should return null for non-Task-annotated declarations', () {
-      var annotated =
-          discoveryGood.discoverDeclaration(goodLib.declarations[#nonTask], {});
+      var annotated = discoveryGood
+          .discoverDeclaration(goodLib.declarations[#nonTask]!, {});
       expect(annotated, isNull);
     });
 
     test('should throw when variable task is null', () {
-      AnnotatedTask f() =>
-          discoveryBad.discoverDeclaration(badLib.declarations[#nullTask], {});
+      AnnotatedTask? f() =>
+          discoveryBad.discoverDeclaration(badLib.declarations[#nullTask]!, {});
       expect(f, throwsA(isA<GrinderException>()));
     });
 
     test('should throw when task getter returns null', () {
-      AnnotatedTask f() => discoveryBad
-          .discoverDeclaration(badLib.declarations[#nullReturningGetter], {});
+      AnnotatedTask? f() => discoveryBad
+          .discoverDeclaration(badLib.declarations[#nullReturningGetter]!, {});
       expect(f, throwsA(isA<GrinderException>()));
     });
 
     test('should throw when task is wront type of declaration', () {
-      AnnotatedTask f() =>
-          discoveryBad.discoverDeclaration(badLib.declarations[#Class], {});
+      AnnotatedTask? f() =>
+          discoveryBad.discoverDeclaration(badLib.declarations[#Class]!, {});
       expect(f, throwsA(isA<GrinderException>()));
     });
 
     test('should throw when depending on non-exported task', () {
-      AnnotatedTask f() => discoveryBad
-          .discoverDeclaration(badLib.declarations[#dependsNonExported], {});
+      AnnotatedTask? f() => discoveryBad
+          .discoverDeclaration(badLib.declarations[#dependsNonExported]!, {});
       expect(f, throwsA(isA<GrinderException>()));
     });
 
     test('should throw when recursively depending on non-exported task', () {
-      AnnotatedTask f() => discoveryBad.discoverDeclaration(
-          badLib.declarations[#recursivelyDependsNonExported], {});
+      AnnotatedTask? f() => discoveryBad.discoverDeclaration(
+          badLib.declarations[#recursivelyDependsNonExported]!, {});
       expect(f, throwsA(isA<GrinderException>()));
     });
 
     test('should throw when depending on invalid task', () {
-      AnnotatedTask f() => discoveryBad
-          .discoverDeclaration(badLib.declarations[#dependsNonTask], {});
+      AnnotatedTask? f() => discoveryBad
+          .discoverDeclaration(badLib.declarations[#dependsNonTask]!, {});
       expect(f, throwsA(isA<GrinderException>()));
     });
 
     test('should throw when annotated with Depends but not Task', () {
-      AnnotatedTask f() => discoveryBad
-          .discoverDeclaration(badLib.declarations[#dependsWithoutTask], {});
+      AnnotatedTask? f() => discoveryBad
+          .discoverDeclaration(badLib.declarations[#dependsWithoutTask]!, {});
       expect(f, throwsA(isA<GrinderException>()));
     });
   });
 
   group('execute returns value', () {
     grinderTest('task from method', () {}, (MockGrinderContext ctx) {
-      var annotated =
-          discoveryGood.discoverDeclaration(goodLib.declarations[#method], {});
+      var annotated = discoveryGood
+          .discoverDeclaration(goodLib.declarations[#method]!, {})!;
       expect(annotated.task.execute(ctx, null), 'someValue');
     });
 
     grinderTest('task from method no-context', () {}, (MockGrinderContext ctx) {
       var annotated = discoveryGood
-          .discoverDeclaration(goodLib.declarations[#noContext], {});
+          .discoverDeclaration(goodLib.declarations[#noContext]!, {})!;
       expect(annotated.task.execute(ctx, null), 'someValue');
     });
 
     grinderTest('task from variable', () {}, (MockGrinderContext ctx) {
       var annotated = discoveryGood
-          .discoverDeclaration(goodLib.declarations[#variable], {});
+          .discoverDeclaration(goodLib.declarations[#variable]!, {})!;
       expect(annotated.task.execute(ctx, null), 'someValue');
     });
 
     grinderTest('task from getter', () {}, (MockGrinderContext ctx) {
-      var annotated =
-          discoveryGood.discoverDeclaration(goodLib.declarations[#getter], {});
+      var annotated = discoveryGood
+          .discoverDeclaration(goodLib.declarations[#getter]!, {})!;
       expect(annotated.task.execute(ctx, null), 'someValue');
     });
   });
