@@ -37,7 +37,7 @@ class GrinderContext {
 
   /// Halt task execution; throws an exception with the given error message.
   @alwaysThrows
-  Null fail(String message) {
+  Never fail(String message) {
     log(grinder.ansi.emphasized('failed: ') + grinder.ansi.error(message));
     throw GrinderException(message);
   }
@@ -56,4 +56,4 @@ void log(String message) => context.log(message);
 
 /// Halt task execution; throws an exception with the given error message.
 @alwaysThrows
-Null fail(String message) => context.fail(message);
+Never fail(String message) => context.fail(message);
